Demonstrate use of swagger in the met-api
Summary:
Add basic use of swagger doc to the api.
To keep the patch small, I will add
swagger-ui in the next patch.

The source in swagger-core/samples/scala-play2
was used as a base.

Test Plan:
Start server by doing "activator run", and the visit

http://localhost:9000/v1/api-docs
http://localhost:9000/v1/api-docs/helloWorld
http://localhost:9000/v1/api-docs/observations

The observations-api is just to have a little bit
realistic api, but it have no tests or correctness
because most of it is going to be thrown away. The
POST is not even implemented. Only documented.
Versioning is not considered.

Reviewers: #continuous_deployment_for_bora, michaeloa

Reviewed By: #continuous_deployment_for_bora, michaeloa

Subscribers: michaeloa, #continuous_deployment_for_bora

Differential Revision: https://phab.met.no/D241