Get rid of the lint warning "If block needs braces".
Summary:
Get rid of the lint warning "If block needs braces". This warning is issued from
this code

if( pred ) foo
else bar

but this is common scala code and conforms to the styleguide.

Test Plan:
arc lint --everything

There should be no lint warning from the scalastyle linter on the form
"If block needs braces".

Reviewers: joa, michaeloa, vegardb

Reviewed By: michaeloa, vegardb

Differential Revision: https://phab.met.no/D430