Extended secureHello test
Summary:
SecureHello now returns different results depending on the method used to access it. Example
and test of how we can filter Basic Auth from OAuth2.

When we implement this for real, we can create the function required as a util function, plug it in here, and test that it works very easily.

Test Plan:
- Access /tests/secureHello using Basic Auth. See the standard Hello message
- Access /tests/secureHello using OAuth2. Verify that it returns the "very securely" message

Reviewers: #met_api, borgem

Reviewed By: #met_api, borgem

Differential Revision: https://phab.met.no/D662